Removed class structure in favor of appending meshes to list
The class structure had some overhead because of naming each class instance so I switched it over to a function that returns the mesh object to a list. I also got all the pages loading different mesh files. As a temporary measure I made the mesh locations random(I will change it to pull from a json file later so each page has all meshes placed in a reasonable way)