General Auto json creation done
Auto generation of routes.json working. will need to make some special case stuff for the homepage so that it randomly selects glb files from projects to display.
working on the gitclone to glb system. folder copying complete, need to convert to glb files then delete .stl in the gitclone-glb.